+++ plugins/ktor-server/krosstalk-ktor-server/src/main/kotlin/com/rnett/krosstalk/ktor/server/Server.kt
@@ -62,6 +62,7 @@ internal fun wrapScopes(route: Route, optional: Boolean, remaining: List<KtorSer
     wrapScopesHelper(route, optional, remaining.toMutableList().asReversed(), final)
 
 public interface KtorKrosstalkServer : KrosstalkServer<KtorServerScope<*>> {
+    @OptIn(KrosstalkPluginApi::class)
     override val server: KtorServer
 }